Adjust Weapon Modifiers
Added 3 modifiers to huge weapons, and reduced short weapons to 1 modifier. The bother of having 2h and the lore of the weapons in question felt like they should have 3 modifiers. And with all the advantages and covertness of small weapons and the lore they seemed to be more like one modifier.